Move kubernetes/kubectl/kind code to internal project layout (#1092)
This is mainly a cleanup to keep all internal related code that could be
reusable in our `internal` directory layout.

Next steps would be to take what's in `internal/kubectl` and merge with this code.
